feature/fix_data_utf8

+++ apiDATN/src/main/kotlin/com/example/apidatn/dto/NewPostDto.kt
@@ -15,7 +15,7 @@ data class NewPostDto (
         var avatar:String?=null,
         var description:String?=null,
         var amountProduct:Int?=null,
-        var priceProduct :Int?=null,
-        var priceDeposit:Int?=null,
+        var priceProduct :Float?=null,
+        var priceDeposit:Float?=null,
         var listImage:MutableList<ImageDto>?=null
     )
